pyproject.toml
[tool.poetry]
name = "oci-ai-code-academy"
version = "0.1.0"
description = "This repository includes code related to the OCI Generative AI Certification Course"
license = "Apache-2.0"
authors = ["Quentin Reul <Quentin.H.Reul@gmail.com>"]
readme = "README.md"
keywords = ["oracle", "oci ai", "generative ai"]

[project.urls]
"Source Code" = "https://github.com/qhreul/oci-ai-code-academy"
"Bug Tracker" = "https://github.com/qhreul/oci-ai-code-academy/issues"

[tool.poetry.dependencies]
chromadb = "^0.5"
llama-index = "^0.10"
llama-index-embeddings-oci-genai = "^0.1"
llama-index-llms-oci-genai = "^0.1"
llama-index-vector-stores-chroma = "^0.1"
oci = "^2.129"
pydantic = "^2.8"
pypdf = "^4.2"
python = "^3.11"
python-dotenv = "^1"
streamlit = "^1.36"

[tool.poetry.group.test.dependencies]
coverage = "*"
mock = "*"
parameterized = "*"
pytest = "^7"
pytest-mock = "*"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[tool.coverage.run]
omit = [
    "tests/*"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
